requirements.txt
# Pacotes necessários para o Bot Framework SDK
botbuilder-core
botbuilder-schema
aiohttp